lambda/requirements.txt
ask-sdk-core==1.19.0
ask-sdk-model==1.82.0
ask-sdk-runtime==1.19.0
openai==0.28.0
python-dotenv==1.0.1
urllib3<2.0
